tidy up lifted file
