+++ pkg/resourceinterpreter/defaultinterpreter/retain.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/karmada-io/karmada/pkg/util"
 	"github.com/karmada-io/karmada/pkg/util/helper"
+	"github.com/karmada-io/karmada/pkg/util/lifted"
 )
 
 // retentionInterpreter is the function that retains values from "observed" object.
@@ -18,23 +19,12 @@ type retentionInterpreter func(desired *unstructured.Unstructured, observed *uns
 func getAllDefaultRetentionInterpreter() map[schema.GroupVersionKind]retentionInterpreter {
 	s := make(map[schema.GroupVersionKind]retentionInterpreter)
 	s[corev1.SchemeGroupVersion.WithKind(util.PodKind)] = retainPodFields
-	s[corev1.SchemeGroupVersion.WithKind(util.ServiceKind)] = retainServiceFields
-	s[corev1.SchemeGroupVersion.WithKind(util.ServiceAccountKind)] = retainServiceAccountFields
+	s[corev1.SchemeGroupVersion.WithKind(util.ServiceKind)] = lifted.RetainServiceFields
+	s[corev1.SchemeGroupVersion.WithKind(util.ServiceAccountKind)] = lifted.RetainServiceAccountFields
 	s[corev1.SchemeGroupVersion.WithKind(util.PersistentVolumeClaimKind)] = retainPersistentVolumeClaimFields
 	s[batchv1.SchemeGroupVersion.WithKind(util.JobKind)] = retainJobSelectorFields
 	return s
 }
-
-/*
-This code is directly lifted from the kubefed codebase. It's a list of functions to update the desired object with values retained
-from the cluster object.
-For reference: https://github.com/kubernetes-sigs/kubefed/blob/master/pkg/controller/sync/dispatch/retain.go#L27-L133
-*/
-
-const (
-	// SecretsField indicates the 'secrets' field of a service account
-	SecretsField = "secrets"
-)
 
 func retainPodFields(desired, observed *unstructured.Unstructured) (*unstructured.Unstructured, error) {
 	desiredPod, err := helper.ConvertToPod(desired)
@@ -66,80 +56,6 @@ func retainPodFields(desired, observed *unstructured.Unstructured) (*unstructure
 	}
 
 	return unstructuredObj, nil
-}
-
-// retainServiceFields updates the desired service object with values retained from the cluster object.
-func retainServiceFields(desired, observed *unstructured.Unstructured) (*unstructured.Unstructured, error) {
-	// healthCheckNodePort is allocated by APIServer and unchangeable, so it should be retained while updating
-	if err := retainServiceHealthCheckNodePort(desired, observed); err != nil {
-		return nil, err
-	}
-
-	// ClusterIP is allocated to Service by cluster, so retain the same if any while updating
-	if err := retainServiceClusterIP(desired, observed); err != nil {
-		return nil, err
-	}
-
-	return desired, nil
-}
-
-func retainServiceHealthCheckNodePort(desired, observed *unstructured.Unstructured) error {
-	healthCheckNodePort, ok, err := unstructured.NestedInt64(observed.Object, "spec", "healthCheckNodePort")
-	if err != nil {
-		return fmt.Errorf("error retrieving healthCheckNodePort from service: %w", err)
-	}
-	if ok && healthCheckNodePort > 0 {
-		if err = unstructured.SetNestedField(desired.Object, healthCheckNodePort, "spec", "healthCheckNodePort"); err != nil {
-			return fmt.Errorf("error setting healthCheckNodePort for service: %w", err)
-		}
-	}
-	return nil
-}
-
-func retainServiceClusterIP(desired, observed *unstructured.Unstructured) error {
-	clusterIP, ok, err := unstructured.NestedString(observed.Object, "spec", "clusterIP")
-	if err != nil {
-		return fmt.Errorf("error retrieving clusterIP from cluster service: %w", err)
-	}
-	// !ok could indicate that a cluster ip was not assigned
-	if ok && clusterIP != "" {
-		err = unstructured.SetNestedField(desired.Object, clusterIP, "spec", "clusterIP")
-		if err != nil {
-			return fmt.Errorf("error setting clusterIP for service: %w", err)
-		}
-	}
-	return nil
-}
-
-// retainServiceAccountFields retains the 'secrets' field of a service account
-// if the desired representation does not include a value for the field.  This
-// ensures that the sync controller doesn't continually clear a generated
-// secret from a service account, prompting continual regeneration by the
-// service account controller in the member cluster.
-func retainServiceAccountFields(desired, observed *unstructured.Unstructured) (*unstructured.Unstructured, error) {
-	// Check whether the secrets field is populated in the desired object.
-	desiredSecrets, ok, err := unstructured.NestedSlice(desired.Object, SecretsField)
-	if err != nil {
-		return nil, fmt.Errorf("error retrieving secrets from desired service account: %w", err)
-	}
-	if ok && len(desiredSecrets) > 0 {
-		// Field is populated, so an update to the target resource does not
-		// risk triggering a race with the service account controller.
-		return desired, nil
-	}
-
-	// Retrieve the secrets from the cluster object and retain them.
-	secrets, ok, err := unstructured.NestedSlice(observed.Object, SecretsField)
-	if err != nil {
-		return nil, fmt.Errorf("error retrieving secrets from service account: %w", err)
-	}
-	if ok && len(secrets) > 0 {
-		err := unstructured.SetNestedField(desired.Object, secrets, SecretsField)
-		if err != nil {
-			return nil, fmt.Errorf("error setting secrets for service account: %w", err)
-		}
-	}
-	return desired, nil
 }
 
 func retainPersistentVolumeClaimFields(desired, observed *unstructured.Unstructured) (*unstructured.Unstructured, error) {
