+++ pkg/util/lifted/discovery.go
@@ -1,4 +1,4 @@
-package detector
+package lifted
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
