+++ pkg/util/objectwatcher/objectwatcher.go
@@ -3,8 +3,6 @@ package objectwatcher
 import (
 	"context"
 	"fmt"
-	"reflect"
-	"strings"
 	"sync"
 
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -19,12 +17,8 @@ import (
 	workv1alpha2 "github.com/karmada-io/karmada/pkg/apis/work/v1alpha2"
 	"github.com/karmada-io/karmada/pkg/resourceinterpreter"
 	"github.com/karmada-io/karmada/pkg/util"
+	"github.com/karmada-io/karmada/pkg/util/lifted"
 	"github.com/karmada-io/karmada/pkg/util/restmapper"
-)
-
-const (
-	generationPrefix      = "gen:"
-	resourceVersionPrefix = "rv:"
 )
 
 // ObjectWatcher manages operations for object dispatched to member clusters.
@@ -213,7 +207,7 @@ func (o *objectWatcherImpl) genObjectKey(obj *unstructured.Unstructured) string 
 
 // recordVersion will add or update resource version records
 func (o *objectWatcherImpl) recordVersion(clusterObj *unstructured.Unstructured, clusterName string) {
-	objVersion := objectVersion(clusterObj)
+	objVersion := lifted.ObjectVersion(clusterObj)
 	objectKey := o.genObjectKey(clusterObj)
 	if o.isClusterVersionRecordExist(clusterName) {
 		o.updateVersionRecord(clusterName, objectKey, objVersion)
@@ -270,55 +264,5 @@ func (o *objectWatcherImpl) NeedsUpdate(clusterName string, desiredObj, clusterO
 		return false, fmt.Errorf("failed to update resource(kind=%s, %s/%s) in cluster %s for the version record does not exist", desiredObj.GetKind(), desiredObj.GetNamespace(), desiredObj.GetName(), clusterName)
 	}
 
-	return objectNeedsUpdate(desiredObj, clusterObj, version), nil
-}
-
-/*
-This code is lifted from the kubefed codebase. It's a list of functions to determine whether the provided cluster
-object needs to be updated according to the desired object and the recorded version.
-For reference: https://github.com/kubernetes-sigs/kubefed/blob/master/pkg/controller/util/propagatedversion.go#L30-L59
-*/
-
-// objectVersion retrieves the field type-prefixed value used for
-// determining currency of the given cluster object.
-func objectVersion(clusterObj *unstructured.Unstructured) string {
-	generation := clusterObj.GetGeneration()
-	if generation != 0 {
-		return fmt.Sprintf("%s%d", generationPrefix, generation)
-	}
-	return fmt.Sprintf("%s%s", resourceVersionPrefix, clusterObj.GetResourceVersion())
-}
-
-// objectNeedsUpdate determines whether the 2 objects provided cluster
-// object needs to be updated according to the desired object and the
-// recorded version.
-func objectNeedsUpdate(desiredObj, clusterObj *unstructured.Unstructured, recordedVersion string) bool {
-	targetVersion := objectVersion(clusterObj)
-
-	if recordedVersion != targetVersion {
-		return true
-	}
-
-	// If versions match and the version is sourced from the
-	// generation field, a further check of metadata equivalency is
-	// required.
-	return strings.HasPrefix(targetVersion, generationPrefix) && !objectMetaObjEquivalent(desiredObj, clusterObj)
-}
-
-// objectMetaObjEquivalent checks if cluster-independent, user provided data in two given ObjectMeta are equal. If in
-// the future the ObjectMeta structure is expanded then any field that is not populated
-// by the api server should be included here.
-func objectMetaObjEquivalent(a, b metav1.Object) bool {
-	if a.GetName() != b.GetName() {
-		return false
-	}
-	if a.GetNamespace() != b.GetNamespace() {
-		return false
-	}
-	aLabels := a.GetLabels()
-	bLabels := b.GetLabels()
-	if !reflect.DeepEqual(aLabels, bLabels) && (len(aLabels) != 0 || len(bLabels) != 0) {
-		return false
-	}
-	return true
+	return lifted.ObjectNeedsUpdate(desiredObj, clusterObj, version), nil
 }
