--- pkg/util/resourcehelper/resourcehelper.go
@@ -1,61 +1,0 @@
-package resourcehelper
-
-import (
-	"fmt"
-	"strings"
-
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/validation"
-)
-
-// This code is directly lifted from the Kubernetes codebase in order to avoid relying on the k8s.io/kubernetes package.
-// For reference: https://github.com/kubernetes/kubernetes/blob/release-1.22/pkg/apis/core/v1/helper/helpers.go,
-// https://github.com/kubernetes/kubernetes/blob/release-1.22/pkg/scheduler/util/utils.go#L144
-
-// IsExtendedResourceName returns true if:
-// 1. the resource name is not in the default namespace;
-// 2. resource name does not have "requests." prefix,
-// to avoid confusion with the convention in quota
-// 3. it satisfies the rules in IsQualifiedName() after converted into quota resource name
-func IsExtendedResourceName(name corev1.ResourceName) bool {
-	if IsNativeResource(name) || strings.HasPrefix(string(name), corev1.DefaultResourceRequestsPrefix) {
-		return false
-	}
-	// Ensure it satisfies the rules in IsQualifiedName() after converted into quota resource name
-	nameForQuota := fmt.Sprintf("%s%s", corev1.DefaultResourceRequestsPrefix, string(name))
-	if errs := validation.IsQualifiedName(string(nameForQuota)); len(errs) != 0 {
-		return false
-	}
-	return true
-}
-
-// IsPrefixedNativeResource returns true if the resource name is in the
-// *kubernetes.io/ namespace.
-func IsPrefixedNativeResource(name corev1.ResourceName) bool {
-	return strings.Contains(string(name), corev1.ResourceDefaultNamespacePrefix)
-}
-
-// IsNativeResource returns true if the resource name is in the
-// *kubernetes.io/ namespace. Partially-qualified (unprefixed) names are
-// implicitly in the kubernetes.io/ namespace.
-func IsNativeResource(name corev1.ResourceName) bool {
-	return !strings.Contains(string(name), "/") ||
-		IsPrefixedNativeResource(name)
-}
-
-// IsHugePageResourceName returns true if the resource name has the huge page
-// resource prefix.
-func IsHugePageResourceName(name corev1.ResourceName) bool {
-	return strings.HasPrefix(string(name), corev1.ResourceHugePagesPrefix)
-}
-
-// IsAttachableVolumeResourceName returns true when the resource name is prefixed in attachable volume
-func IsAttachableVolumeResourceName(name corev1.ResourceName) bool {
-	return strings.HasPrefix(string(name), corev1.ResourceAttachableVolumesPrefix)
-}
-
-// IsScalarResourceName validates the resource for Extended, Hugepages, Native and AttachableVolume resources
-func IsScalarResourceName(name corev1.ResourceName) bool {
-	return IsExtendedResourceName(name) || IsHugePageResourceName(name) ||
-		IsPrefixedNativeResource(name) || IsAttachableVolumeResourceName(name)
-}
